pyproject.toml
[tool.poetry]
name = "deira"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
dagster = "^0.14.20"
dagit = "^0.14.20"
boto3 = "^1.24.7"
pandas = "^1.4.2"
install = "^1.3.5"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
